UI: Status indicator border around program scene
Add a new general setting to enable, and control the size of an
indicator border that appears around the scene display when
streaming or recording. In studio mode this only appears on the
Program scene. This gives users a new more obvious at a glance
indicator that they are currently streaming or recording. The border
color can be assigned through new accessibility colors.